feat(net-hook): reusable load hook library with TLS verify + cache
Extract the reference image/content load hook out of the per-consumer
`test_image_hook.h` blob and into a proper static library under
`mcc/net_hook/`. Host applications now link `libhtmlview_nethook.a` and
call `HTMLviewNet_InitHook(&hook)`; the hook opens bsdsocket / AmiSSL
lazily from the decoder task so hosts no longer need any OpenLibrary
plumbing of their own.

Adds three opt-in runtime knobs on top of the extraction:

- `HTMLviewNet_SetCABundle` / `SetVerifyMode` -- flip AmiSSL off the
  previous `SSL_VERIFY_NONE` default to auto-discovered peer verify
  (checks `AmiSSL:Certs/curl-ca-bundle.crt` and ENVARC mirrors),
  including hostname matching via `SSL_set1_host`.
- `HTMLviewNet_SetCacheDir` / `SetCacheTTL` -- simple on-disk response
  cache keyed by FNV-1a-64 of the URL, storing `<key>.body` + `.meta`
  pairs. Disabled by default; caps entries at 8 MB, non-chunked 200
  responses only. ETag revalidation is deferred to a follow-up.

Also fixes a latent CI break in `mcc/Makefile`: the `\$(AMISSL_SDK_READY)`
rule was defined before `all:`, so a bare `make OS=os3` used the SDK
fetch as the default goal and exited 0 without building anything. The
13.6 release job hit exactly this and silently uploaded an artifact
with no `.mcc` inside. Pinning `.DEFAULT_GOAL := all` at the top of the
file makes `make` build by default; the new `libhtmlview_nethook.a`
target and net_hook obj subdir are wired in alongside.

The full improvement plan / session handoff lives in IMPROVEMENTS.md.
